backend: Resolve issues with websocket token encoding
This resolves an issue introduced in #3142 when decoding the bearer
token from the websocket Sec-Websocket-Protoccol it assumes the
token is padded.

When testing with tokens generated for Azure AKS and GKE, I noticed
headlamp started to fail all websocket connections and upon further
investigation, the tokens that did not have padding (==) were not
decoded properly and the authorization header was malformed when sent
to K8s.

This attempts to decode the token additionally with RawStdEncoding
to catch these edge cases.

Signed-off-by: David Dobmeier <david.dobmeier@live.com>
Signed-off-by: David Dobmeier <david.dobmeier@dcsg.com>