backend: Implement support for Azure AKS managed OIDC auth
This is one option that is more flexible for other auth providers
and allows overriding the client id and issuer for validation
as well as allows for grabbing the access_token instead of
id_token which is required for AKS's implementation of OIDC
managed auth.

Alternatively this can be implemented with a single flag
specifically for Azure AKS managed OIDC (which I will take
a stab at implementing before submitting a PR) to have both
as options.

Signed-off-by: David Dobmeier <david.dobmeier@live.com>