Update helm chart for oidc validator override support
Updates the chart with the new parameters + also updates the tests.

Fully backwards compatible with existing installations. For users using
`external-secret` for oidc config will need to specify the useAccessToken
in the values yaml or set `- "-oidc-use-access-token=$(OIDC_USE_ACCESS_TOKEN)"`
in the extraArgs. This is done to maintain backwards compatibility.

Signed-off-by: David Dobmeier <david.dobmeier@dcsg.com>