Add documentation for Azure AKS OIDC with Entra ID
Adds documentation for the new parameters for token validation, using
`access_token` instead of `id_token`, and includes a full walkthrough
of setting up an AKS cluster with OIDC + deploying headlamp using it.

Credit to `magohl` for some of the photos and steps used in documentation,
linked to in `Issue 2480`.

Signed-off-by: David Dobmeier <david.dobmeier@dcsg.com>