backend: auth: Extract GetExpiryUnixTimeUTC from headlamp.go
Explain the type of time in the function name to make it
a bit clearer.

The exp field is in UTC, so we return UTC. Generally all server
side code should be using UTC rather than converting to a
local time zone.

Add some documentation and tests with edge cases.