frontend: runCommand: Use permissionSecrets and private desktopApi funcs
The runCommand now takes permission secrets, and desktop API
functions so that it can securely communicate without other plugins
being able to snoop on the permission secrets required by the app
to run different commands.

These secrets and functions will be passed into this function by
a future change to runPlugin which hides the secrets and desktopAPI
functions from the rest of the plugins.