frontend: Fix portforwarding to a pod when not in a cluster
When opening a pod, navigating to a different or to a no-cluster view,
and attempting to portforwarding that pod, this action was failing
because the cluster context was taken from whatever was in the URL at
that moment and that was none. This patch fixes that by relying on the
resource's cluster or the initial one.

Signed-off-by: Joaquim Rocha <joaquim.rocha@microsoft.com>