app: runCmd: Fix issue on windows/mac with file URLs and import
The import() doesn't work properly with file:// URLs.

Also it seems the pathToFileURL is buggy on windows, it gives a
URL with file:/// instead of file://

Also Mac with full paths doesn't work with file:// URLs.

The .. and / checking were redundant. What really matters is that
the scripts are loaded from the plugin folder.