frontend: app: Add faster dev mode alternative with rsbuild
This adds an alternative dev mode to app and frontend:
```ts
npm run star
```

It leaves the previous "npm start" there, and leaves the build
and test systems as before.

This uses rsbuild which is a webpack compatible build tool that
does not send 850 files in dev mode.

Rather than minutes or 10s of seconds it is closer to the speed
of the compiled app or the previous webpack based dev mode.

- `cd frontend/ && npm run star`
- `cd app/ && npm run star`
  - compare app speed of dev mode on a slow windows computer in cmd
  - even on a fast new mac laptop, it is seconds quicker
- This still works: `cd frontend/ && npm run start`
- win app still works `make frontend backend app-win`
- mac app `make frontend backend app-mac`
- container e2e tests still work